Fix -Wodr violations in SQLite related drivers
