+++ gdal/ogr/ogrsf_frmts/gpkg/makefile.vc
@@ -7,14 +7,10 @@ GDAL_ROOT	=	..\..\..
 
 !INCLUDE $(GDAL_ROOT)\nmake.opt
 
-EXTRAFLAGS = -I.. -I..\.. -I..\sqlite  -I..\..\..\frmts\mem -I..\..\..\frmts $(SQLITE_INC) $(SQLITE_HAS_COLUMN_METADATA_EXTRAFLAGS) $(SPATIALITE_412_OR_LATER_EXTRAFLAGS)
+EXTRAFLAGS = -I.. -I..\.. -I..\sqlite  -I..\..\..\frmts\mem -I..\..\..\frmts $(SQLITE_INC) $(SQLITE_HAS_COLUMN_METADATA_EXTRAFLAGS)
 
 !IFDEF SQLITE_HAS_COLUMN_METADATA
 SQLITE_HAS_COLUMN_METADATA_EXTRAFLAGS = -DSQLITE_HAS_COLUMN_METADATA
-!ENDIF
-
-!IFDEF SPATIALITE_412_OR_LATER
-SPATIALITE_412_OR_LATER_EXTRAFLAGS = -DSPATIALITE_412_OR_LATER
 !ENDIF
 
 default:	$(OBJ)
